Don't specify a version for Bundler in the gemspec
This is because Circle has an outdated version of Bundler installed and
complains because we're trying to use a newer version. It also uses a
gemset for its gems so it's a pain to override this behavior, so the
easiest thing to do is to just remove the version from the gemspec.