Ensure Brakeman runner is loaded last
There's an issue with Pronto when running the Brakeman and
rails_best_practices runners in a specific order
(https://github.com/mmozuras/pronto/issues/80). If Brakeman is loaded
before rails_best_practices, it can raise false positives, so we should
make sure that Brakeman is loaded last when loading the Pronto runners.